Added kronecker product
