+++ dislib/__init__.py
@@ -1,7 +1,8 @@
 import os
 
-from dislib.data.array import random_array, apply_along_axis, array, \
-    load_svmlight_file, load_txt_file, load_npy_file, zeros
+from dislib.data.array import random_array, apply_along_axis, array, zeros
+from dislib.data.io import load_svmlight_file, load_npy_file, load_txt_file
+from dislib.math import kron
 
 name = "dislib"
 version_file = os.path.join(os.path.dirname(os.path.abspath(__file__)),
@@ -25,4 +26,4 @@ else:
         __version__ = 'unknown'
 
 __all__ = ['load_txt_file', 'load_svmlight_file', 'random_array',
-           'apply_along_axis', 'array', 'load_npy_file', 'zeros']
+           'apply_along_axis', 'array', 'load_npy_file', 'zeros', 'kron']
