+++ dislib/data/__init__.py
@@ -1,5 +1,5 @@
-from dislib.data.array import array, random_array, apply_along_axis, \
-    load_txt_file, load_svmlight_file, load_npy_file, zeros
+from dislib.data.array import array, random_array, apply_along_axis, zeros
+from dislib.data.io import load_txt_file, load_npy_file, load_svmlight_file
 
 __all__ = ['load_txt_file', 'load_svmlight_file', 'array', 'random_array',
            'apply_along_axis', 'load_npy_file', 'zeros']
