+++ dislib/math/__init__.py
@@ -1,0 +1,3 @@
+from dislib.math.base import kron
+
+__all__ = ['kron']
